Feature/file cards (#4)
* plan - horizontally scrollable file cards instead of file-recommendation trees

* Add horizontal scrolling and pagination dots to FileCard component

Enable horizontal scrolling using keyboard and add pagination dots to the FileCard component to indicate the number of files and current position.

* Add file number indicator to FileCard component

Display the file number out of the total number of files on the top right of each FileCard component.

* Show File Path in File Card

Display the file path below the file name in the FileCard component using VS Code color variables.

* Synchronize File Card Scrolling with Editor

Implement synchronized scrolling between file cards in  and the active editor in VS Code.

* renamings

* Fix file card width issue

The file card component in  is exceeding the width of its parent container. This needs to be adjusted so that the card content stays within the bounds of its parent.

* create new plan logic fix

* creation logic for non existent files fixed

* Implement single card view with swipe navigation and dot-based jumping in FormattedPlanPreview

* css

* error handling

* open active plan file in editor

* not more chund adding once whole file is loaded, in case there's some race condition